agent/requirements.txt
openai
google-generativeai
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv
requests==2.31.0
PyPDF2==3.0.1
numpy==1.24.3
sentence-transformers==2.2.2
